Consolidate image session handling.
